+++ iExtra/UI/CornerRadius/CornerRadiusStyle.swift
@@ -1,0 +1,14 @@
+//
+//  BaseViewController.swift
+//  Pinny
+//
+//  Created by Daniel Saidi on 2018-05-09.
+//  Copyright © 2018 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public protocol CornerRadiusStyle {
+    
+    var radius: Int { get }
+}